[fix_edit_remove_permissions]: Удалять/редактировать только себя

+++ server/routes/users.js
@@ -14,6 +14,13 @@ export default (app) => {
             reply.render('users/new', { user });
         })
         .get('/users/:id/edit', { name: 'editUser' }, async (req, reply) => {
+            if (req?.user?.id !== +req.params.id) {
+                req.flash('error', i18next.t('flash.users.view.error'));
+                reply.redirect(app.reverse('users'));
+
+                return reply;
+            }
+
             const user = await app.objection.models.user.query().findById(req.params.id);
             const userInfo = await user.$relatedQuery('usersinfo');
 
@@ -76,6 +83,13 @@ export default (app) => {
             return reply;
         })
         .delete('/users/:id', { name: 'deleteUser' }, async (req, reply) => {
+            if (req?.user?.id !== +req.params.id) {
+                req.flash('error', i18next.t('flash.users.view.error'));
+                reply.redirect(app.reverse('users'));
+
+                return reply;
+            }
+
             try {
                 await app.objection.models.userInfo.query().delete().where('user', +req.params.id);
                 await app.objection.models.user.query().deleteById(+req.params.id);
